requirement.txt
streamlit
pillow
numpy
scikit-learn
joblib
matplotlib
seaborn
pandas
ultralytics
shap
scikit-image